pyproject.toml
[project]
name = "job-hunter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "ruff>=0.12.7",
    "python-jobspy>=1.1.82",
    "uvicorn[standard]>=0.20.0",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.10.1",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "greenlet>=3.2.3",
]
